Add content type checking to uploads:
To limit files types allowed to be uploaded to image files, server.UploadHandler
now checks the content type of the file using the first chunk received.
If the chunk doesn't match an image/ content type, the connection is closed
with a response code 400.

No checking is done on the content-type header, but it might worth
checking for image/ or multipart/form-date types. (But does this increase
security?)