RemoveHandler changes
Prior to this commit, passing a name parameter that included directories
before the filename would remove the file in the watched directory if a
file with the same name existed.
For example, if the watched directory contains a file named image.jpg,
and the user passes a name parameter of directory/name/image1.jpg, the
image would still be removed.

The fix for this checks the result of filepath.Dir. If the result is not
".", indicating there is no directory path before the file name, the
HTTP request is returned with code 400.

The HTTP request now must also be a DELETE type.