Fixes #1
Reading and writing chunks of multi-part requests straight
to file caused many filesystem events to be fired, causing
an issue for the imageservice if it tried to load the image
file before it was finished being written.

Storing the whole image in memory then writing it in one
go mitigates this issue. One issue could be the amount of
memory images consume when being uploaded, but since
uploads are short lived as long as resources are released
correctly it shouldn't be an issue.